@@ -17,9 +17,16 @@ limitations under the License.
 package enricher
 
 import (
+	"bufio"
+	"bytes"
 	"context"
+	"fmt"
+	"io"
+	"io/fs"
 	"net"
 	"os"
+	"path/filepath"
+	"strings"
 
 	"github.com/jellydator/ttlcache/v3"
 	"github.com/nxadm/tail"
@@ -30,12 +37,21 @@ import (
 	"k8s.io/client-go/rest"
 
 	api "sigs.k8s.io/security-profiles-operator/api/grpc/metrics"
+	"sigs.k8s.io/security-profiles-operator/internal/pkg/daemon/enricher/auditsource"
 	"sigs.k8s.io/security-profiles-operator/internal/pkg/daemon/enricher/types"
 	"sigs.k8s.io/security-profiles-operator/internal/pkg/daemon/metrics"
 	"sigs.k8s.io/security-profiles-operator/internal/pkg/util"
 )
 
-type defaultImpl struct{}
+type defaultImpl struct {
+	fsys fs.FS // Must be initialized by newDefaultImpl
+}
+
+func newDefaultImpl() *defaultImpl {
+	return &defaultImpl{
+		fsys: os.DirFS("/"),
+	}
+}
 
 //go:generate go run github.com/maxbrunsfeld/counterfeiter/v6 -generate -header ../../../../hack/boilerplate/boilerplate.generatego.txt
 //counterfeiter:generate . impl
@@ -43,6 +59,8 @@ type impl interface {
 	Getenv(key string) string
 	Dial() (*grpc.ClientConn, context.CancelFunc, error)
 	Close(*grpc.ClientConn) error
+	StartTail(src auditsource.AuditLineSource) (chan *types.AuditLine, error)
+	TailErr(src auditsource.AuditLineSource) error
 	TailFile(filename string, config tail.Config) (*tail.Tail, error)
 	Lines(tailFile *tail.Tail) chan *tail.Line
 	Reason(tailFile *tail.Tail) error
@@ -60,6 +78,9 @@ type impl interface {
 	Chown(string, int, int) error
 	Stat(string) (os.FileInfo, error)
 	RemoveAll(string) error
+	CmdlineForPID(pid int) (string, error)
+	PrintJsonOutput(w io.Writer, output string)
+	EnvForPid(pid int) (map[string]string, error)
 }
 
 func (d *defaultImpl) Getenv(key string) string {
@@ -72,6 +93,14 @@ func (d *defaultImpl) Dial() (*grpc.ClientConn, context.CancelFunc, error) {
 
 func (d *defaultImpl) Close(conn *grpc.ClientConn) error {
 	return conn.Close()
+}
+
+func (d *defaultImpl) StartTail(src auditsource.AuditLineSource) (chan *types.AuditLine, error) {
+	return src.StartTail()
+}
+
+func (d *defaultImpl) TailErr(src auditsource.AuditLineSource) error {
+	return src.TailErr()
 }
 
 func (d *defaultImpl) TailFile(
@@ -164,4 +193,81 @@ func (d *defaultImpl) Stat(name string) (os.FileInfo, error) {
 
 func (d *defaultImpl) RemoveAll(path string) error {
 	return os.RemoveAll(path)
+}
+
+func (d *defaultImpl) CmdlineForPID(
+	pid int,
+) (string, error) {
+	var retErr error
+
+	cmdline := fmt.Sprintf("proc/%d/cmdline", pid)
+
+	file, err := d.fsys.Open(filepath.Clean(cmdline))
+	if err != nil {
+		retErr = fmt.Errorf("%w: %w", ErrProcessNotFound, err)
+
+		return "", retErr
+	}
+
+	defer func() {
+		cerr := file.Close()
+		if retErr == nil {
+			retErr = cerr
+		}
+	}()
+
+	var sb strings.Builder
+
+	scanner := bufio.NewScanner(file)
+	for scanner.Scan() {
+		sb.WriteString(strings.ReplaceAll(scanner.Text(), "\u0000", " "))
+	}
+
+	if err := scanner.Err(); err != nil {
+		retErr = fmt.Errorf("%w: %w", ErrCmdlineNotFound, err)
+
+		return "", retErr
+	}
+
+	return sb.String(), retErr
+}
+
+func (d *defaultImpl) EnvForPid(pid int) (map[string]string, error) {
+	var retErr error
+
+	envFile := fmt.Sprintf("proc/%d/environ", pid)
+	envMap := make(map[string]string)
+
+	content, err := fs.ReadFile(d.fsys, filepath.Clean(envFile))
+	if err != nil {
+		retErr = fmt.Errorf("%w: %w", ErrProcessNotFound, err)
+
+		return envMap, retErr
+	}
+
+	envVars := bytes.Split(content, []byte{0})
+
+	for _, envVarBytes := range envVars {
+		envVar := string(envVarBytes)
+		if envVar == "" {
+			continue
+		}
+
+		// Ignore keys with no values
+		parts := strings.SplitN(envVar, "=", 2)
+		if len(parts) == 2 {
+			key := parts[0]
+			value := parts[1]
+			envMap[key] = value
+		}
+	}
+
+	return envMap, nil
+}
+
+func (d *defaultImpl) PrintJsonOutput(w io.Writer, output string) {
+	_, err := fmt.Fprintln(w, output)
+	if err != nil {
+		fmt.Printf("error printing json output: %v", err)
+	}
 }
